requirements2.txt
dotenv
streamlit
google-generativeai